move partition key field extraction to partition resolver
